fix state _cleanup and unittest
