+++ couler/core/states.py
@@ -71,7 +71,7 @@ def _cleanup():
     """Cleanup the cached fields, just used for unit test.
     """
     global _secrets, _update_steps_lock, _dag_caller_line, _upstream_dag_task, _upstream_dag_depends_logic, workflow, _steps_outputs  # noqa: E501
-    global _exit_handler_enable, _when_prefix, _when_task  # noqa: E501
+    global _exit_handler_enable, _when_prefix, _when_task, _while_steps, _concurrent_func_line  # noqa: E501
     _secrets = {}
     _update_steps_lock = True
     _dag_caller_line = None
@@ -80,5 +80,7 @@ def _cleanup():
     _exit_handler_enable = False
     _when_prefix = None
     _when_task = None
+    _while_steps = OrderedDict()
+    _concurrent_func_line = -1
     _steps_outputs = OrderedDict()
     workflow.cleanup()
